Inline assume to avoid performance regression
The intrinsic assume is called in a lot of hot paths throughout alloc. It should not compiled to any code and just inform optimizations, however the stable polyfill is apparently not inlined enough as can be seen while profiling some some simple vector operations.

This commit improves the performance of a benchmark using Vec a fair amount although it does not quite achieves the performance of using the standard Vec.